Comment out student analytics for now.
Really wanted to see this get into the sprint, but it's better
to get a strong start on this next sprint than it is to try to
rush this in on the last day of this one.

In retrospect, it was a super bad idea to skimp on testing from
the start.