Fix image uploader on multiple uploads (#180)
This was a tricky bug to track down.  The img.onload event was
getting triggered immediately after the cropper was deleted,
which triggered the creation of another cropper.

Leading to a seemingly haunted console in which the DOM elements
I had just deleted and logged as deleted were still there..